# 12_Backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ScottsTechX Commerce OS API",
    "version": "1.0.0",
    "description": "Trust-gated marketplace backend for the Uganda MVP slice.\n\nThis spec is the source of truth for clients (Flutter, web, admin). It mirrors\nthe implementation in `12_Backend/src/server.ts` and the migration-derived\nschema. Endpoints marked `STUB` return HTTP 501 with `X-Stub-Reason` header\nand do NOT perform any external I/O.\n\n## Authentication\n\nAll endpoints require `Authorization: Bearer ***` where the token is an\nHS256 JWT with the following claims:\n  - `iss`: must equal `scottstechx`\n  - `aud`: must equal `scottstechx-api`\n  - `sub`: user's UUID\n  - `role`: one of `buyer | driver | seller | admin`\n  - `exp`: present and not in the past\n\n## Idempotency\n\n`POST` and `PATCH` endpoints that change state require an `Idempotency-Key`\nheader of at least 8 characters. Replays with the same body return the\noriginal response. Replays with a different body under the same key return\nHTTP 409.\n\n## Error format\n\nAll errors return JSON: `{ error: string, message: string, details?: any }`.\nThe `error` field is a stable machine-readable code.",
    "license": { "name": "Proprietary" }
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Local dev" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "DeliveryAddress": {
        "type": "object",
        "required": ["line1", "city", "country"],
        "properties": {
          "line1": { "type": "string", "maxLength": 200 },
          "city": { "type": "string", "maxLength": 100 },
          "country": { "type": "string", "enum": ["UG"] }
        }
      },
      "CheckoutRequest": {
        "type": "object",
        "required": ["items", "delivery_address"],
        "properties": {
          "items": {
            "type": "array", "minItems": 1, "maxItems": 50,
            "items": { "$ref": "#/components/schemas/OrderItemInput" }
          },
          "delivery_address": { "$ref": "#/components/schemas/DeliveryAddress" }
        }
      },
      "OrderItemInput": {
        "type": "object",
        "required": ["product_id", "qty"],
        "properties": {
          "product_id": { "type": "string", "format": "uuid" },
          "qty": { "type": "integer", "minimum": 1, "maximum": 999 }
        }
      },
...
      "PodRequest": {
        "type": "object",
        "required": ["order_id", "action", "gps_lat", "gps_lng"],
        "properties": {
          "order_id": { "type": "string", "format": "uuid" },
          "action": { "type": "string", "enum": ["pickup", "deliver"] },
          "gps_lat": { "type": "number", "minimum": -90, "maximum": 90 },
          "gps_lng": { "type": "number", "minimum": -180, "maximum": 180 },
          "notes": { "type": "string", "maxLength": 500 }
        }
      },
...
      "GoogleAuthRequest": {
        "type": "object",
        "required": ["idToken"],
        "properties": {
          "idToken": { "type": "string", "minLength": 10 },
          "role": { "type": "string", "enum": ["buyer","seller","driver"] }
        }
      },
      "GoogleAuthResponse": {
        "type": "object",
        "required": ["token","userId","role","email"],
        "properties": {
          "token": { "type": "string" },
          "userId": { "type": "string" },
          "role": { "type": "string" },
          "email": { "type": "string" },
          "expiresAt": { "type": "string" },
          "trustTier": { "type": "string" }
        }
      },
...
      "AiSellerSuggestRequest": {
        "type": "object",
        "required": ["type"],
        "properties": {
          "type": { "type": "string", "enum": ["product_description","auto_price","category","inventory_warning"] },
          "draft": { "type": "object" },
          "context": { "type": "object" }
        }
      },
      "AiSellerSuggestResponse": {
        "type": "object",
        "required": ["suggestion","reasoning","confidence","provider"],
        "properties": {
          "suggestion": { "type": "string" },
          "reasoning": { "type": "string" },
          "confidence": { "type": "number" },
          "provider": { "type": "string" }
        }
      },
      "AiCustomerChatRequest": {
        "type": "object",
        "required": ["sessionId","message"],
        "properties": {
          "sessionId": { "type": "string" },
          "message": { "type": "string" },
          "history": { "type": "array", "items": { "type": "object" } }
        }
      },
      "AiCustomerChatResponse": {
        "type": "object",
        "required": ["reply","provider"],
        "properties": {
          "reply": { "type": "string" },
          "provider": { "type": "string" }
        }
      },
      "AiReasonRequest": {
        "type": "object",
        "required": ["sellerId"],
        "properties": {
          "sellerId": { "type": "string", "format": "uuid" },
          "context": { "type": "object" }
        }
      },
      "AiReasonResponse": {
        "type": "object",
        "required": ["trustReasoning","rankReasoning","recommendation","provider","confidence"],
        "properties": {
          "trustReasoning": { "type": "string" },
          "rankReasoning": { "type": "string" },
          "recommendation": { "type": "string" },
          "provider": { "type": "string" },
          "confidence": { "type": "number" }
        }
      },
...
      "PostChatMessageRequest": {
        "type": "object",
        "required": ["sessionId","content"],
        "properties": {
          "sessionId": { "type": "string" },
          "content": { "type": "string" },
          "recipientUserId": { "type": "string", "nullable": true },
          "role": { "type": "string", "enum": ["buyer","seller","ai","system"] }
        }
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/v1/orders/checkout": {
      "post": {
        "tags": ["orders"], "summary": "Initiate a buyer order checkout",
        "parameters": [{ "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "minLength": 8 } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CheckoutRequest" } } } },
        "responses": {
          "201": { "description": "Order created" },
          "200": { "description": "Idempotent replay" },
          "400": { "description": "Validation error" },
          "401": { "description": "Missing or invalid JWT" },
          "403": { "description": "Caller is not a buyer" },
          "409": { "description": "Idempotency-Key reused with different body" },
          "422": { "description": "Stock or validation failure" },
          "501": { "description": "Stub: mobile money not yet integrated" }
        }
      }
    },
    "/api/v1/logistics/pod": {
      "post": {
        "tags": ["logistics"], "summary": "Submit proof of delivery",
        "parameters": [{ "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "minLength": 8 } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PodRequest" } } } },
        "responses": {
          "200": { "description": "POD accepted" },
          "400": { "description": "Validation error" },
          "401": { "description": "Missing or invalid JWT" },
          "403": { "description": "Caller is not the assigned driver" },
          "404": { "description": "Order not found" },
          "409": { "description": "Illegal state transition" }
        }
      }
    },
...
    "/api/v1/chat/messages": {
      "get": {
        "tags": ["marketplace"], "summary": "List chat messages for a session",
        "parameters": [
          { "name": "sessionId", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "since", "in": "query", "required": false, "schema": { "type": "string", "format": "date-time" } }
        ],
        "responses": { "200": { "description": "Messages" } }
      },
      "post": {
        "tags": ["marketplace"], "summary": "Post a chat message",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PostChatMessageRequest" } } } },
        "responses": { "201": { "description": "Created" } }
      }
    },
...
    "/api/v1/ai/seller-suggest": {
      "post": {
        "tags": ["ai"], "summary": "AI suggestion for seller copy/pricing/category/inventory",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AiSellerSuggestRequest" } } } },
        "responses": {
          "200": { "description": "Suggestion", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AiSellerSuggestResponse" } } } },
          "401": { "description": "Missing or invalid JWT" },
          "403": { "description": "Caller is not a seller" },
          "429": { "description": "Rate limited" },
          "503": { "description": "AI not configured (set LLM_API_KEY)" }
        }
      }
    },
    "/api/v1/ai/customer-chat": {
      "post": {
        "tags": ["ai"], "summary": "AI customer chatbot",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AiCustomerChatRequest" } } } },
        "responses": {
          "200": { "description": "Reply", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AiCustomerChatResponse" } } } },
          "401": { "description": "Missing or invalid JWT" },
          "429": { "description": "Rate limited" },
          "503": { "description": "AI not configured (set LLM_API_KEY)" }
        }
      }
    },
    "/api/v1/ai/reason": {
      "post": {
        "tags": ["ai"], "summary": "AI trust/rank reasoning for a seller",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AiReasonRequest" } } } },
        "responses": {
          "200": { "description": "Reasoning", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AiReasonResponse" } } } },
          "401": { "description": "Missing or invalid JWT" },
          "429": { "description": "Rate limited" },
          "503": { "description": "AI not configured (set LLM_API_KEY)" }
        }
      }
    },
    "/api/v1/auth/google": {
      "post": {
        "tags": ["auth"], "summary": "Exchange Google idToken for our own JWT",
        "security": [],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GoogleAuthRequest" } } } },
        "responses": {
          "200": { "description": "JWT issued", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GoogleAuthResponse" } } } },
          "400": { "description": "Validation error" },
          "503": { "description": "Google sign-in not configured (set GOOGLE_CLIENT_ID)" }
        }
      }
    }
  }
}